Move rename/delete dialogs
