+++ application/ui/src/components/project-dialogs/edit-project-name-dialog.component.tsx
@@ -17,18 +17,25 @@ import {
 } from '@geti/ui';
 import { isEmpty } from 'lodash-es';
 
-import { usePatchProject } from '../../../../../hooks/api/project.hook';
+import { usePatchProject } from '../../hooks/api/project.hook';
 
-interface EditProjectNameDialogProps {
+type EditProjectNameDialogProps = {
     onClose: () => void;
     isOpen: boolean;
     projectId: string;
     projectName: string;
-}
+    onSuccess?: (newName: string) => void;
+};
 
 const PROJECT_NAME_MAX_LENGTH = 100;
 
-export const EditProjectNameDialog = ({ onClose, isOpen, projectId, projectName }: EditProjectNameDialogProps) => {
+export const EditProjectNameDialog = ({
+    onClose,
+    isOpen,
+    projectId,
+    projectName,
+    onSuccess,
+}: EditProjectNameDialogProps) => {
     const patchProjectMutation = usePatchProject();
     const [newProjectName, setNewProjectName] = useState(projectName);
 
@@ -46,6 +53,7 @@ export const EditProjectNameDialog = ({ onClose, isOpen, projectId, projectName 
                 onSuccess: () => {
                     onClose();
                     toast({ type: 'success', message: 'Project updated successfully' });
+                    onSuccess?.(newName);
                 },
             }
         );
@@ -58,7 +66,7 @@ export const EditProjectNameDialog = ({ onClose, isOpen, projectId, projectName 
             return;
         }
 
-        await editProjectName(newProjectName);
+        editProjectName(newProjectName);
     };
 
     return (
